refactor(cli): extract sanitizeFilename to shared lib/file-utils
The same sanitizeFilename implementation lived in both bin/commands/attachments.js
and bin/commands/export.js, so any future fix had to be made twice. Move it to
lib/file-utils.js and import from both call sites.
