+++ bin/commands/attachments.js
@@ -2,19 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const chalk = require('chalk');
 const inquirer = require('inquirer');
-
-function sanitizeFilename(filename) {
-  if (!filename || typeof filename !== 'string') {
-    return 'unnamed';
-  }
-  const stripped = path.basename(filename.replace(/\\/g, '/'));
-  const cleaned = stripped
-    // eslint-disable-next-line no-control-regex
-    .replace(/[\\/:*?"<>|\x00-\x1f]/g, '_')
-    .replace(/^\.+/, '')
-    .trim();
-  return cleaned || 'unnamed';
-}
+const { sanitizeFilename } = require('../../lib/file-utils');
 
 function registerAttachmentCommands(program, { withClient }) {
   program
